[feature] Move notification module into root of package
It is no longer part of the GitHub integration process, but will be used on all file operations in the editor.
