+++ packages/prosemirror-lwdita/src/github-integration/request.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-import { showToast } from './toast';
+import { showToast } from '../toast';
 import { clientID, serverURL } from '../app-config';
 import { exchangeOAuthCodeForAccessToken } from './github.plugin';
